# internal/swagger/geofence-openapi.yaml
openapi: 3.0.3
info:
  title: Geofence Service API
  version: 1.0.0
  description: |
    CRUD and spatial lookup APIs for agency/client geofences. Geofences are
    stored server-side as Oracle Spatial MDSYS.SDO_GEOMETRY and scoped by
    agency (mandatory) and client (optional — a fence with no client_id
    applies agency-wide). Every request/response carries an
    x-correlation-id header for distributed tracing/log correlation.
servers:
  - url: /api/v1
paths:
...
  /agencies/{agency_id}/geofences:list:
    parameters:
      - $ref: '#/components/parameters/AgencyId'
      - $ref: '#/components/parameters/CorrelationId'
    post:
      summary: List geofences for an agency.
      operationId: listGeofences
      x-http-method: LIST
      x-actual-path: /agencies/{agency_id}/geofences
      description: >
        **Must be sent as `LIST /agencies/{agency_id}/geofences`, not
        `POST /agencies/{agency_id}/geofences:list`** — the `:list` suffix
        and `post` method are used here purely to give this operation a
        distinct path from `POST /agencies/{agency_id}/geofences` (create a
        geofence), since OpenAPI 3.0 cannot key two operations off the same
        path with the same declared method (`x-http-method: LIST`). The
        composite (agency_id, client_id) index backs this query — it is the
        primary access pattern for the table. Also accepts an optional JSON
        body (search/pagination/data envelope, mirroring sso-user-service's
        list-users request) as an alternative to the query params below —
        when the body has none of `search`/`pagination`/`data`, is empty, or
        isn't JSON, every field falls back to its query param instead. When
        a body *is* used, every query param is ignored except `client_id`,
        which has no body field and is always read from the query string.
        The agency scope is always resolved server-side — the agency_id path
        param, plus (with includeSubagency) the subagencies the caller may
        see — and can only be narrowed further, via home_agency. A 400 is
        returned for an invalid `client_id` / `status` / `sort_by` /
        `includeSubagency` /
        `home_agency` / `exclude_from_colocation` query value, or a
        non-UUID `search.home_agency` entry (`ERR_INVALID_QUERY` et al.,
        naming the offending input), or for a malformed body — including a
        wrong-typed or null-containing `search.home_agency` /
        `search.exclude_from_colocation` (`ERR_BIND_PAYLOAD`, input
        `geofence_req`).
      tags: [Geofences]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeofenceListRequest'
            examples:
              searchByName:
                summary: Search by name
                value:
                  search: { any: 'Downtown' }
              searchByAddress:
                summary: Search by address
                value:
                  search: { any: 'Main St' }
              filterByHomeAgency:
                summary: Only geofences owned by these (sub)agencies
                value:
                  search: { home_agency: ['8396b27b-95ef-45cb-9b4c-1132812a33e1', '5c2e9a4d-1f3b-4d7e-8a6c-2b9f0e1d3c4a'] }
              searchByExcludedFromColocation:
                summary: Only geofences excluded from co-location
                value:
                  search: { any: 'true' }
              fieldFiltersAnded:
                summary: Per-field filters (ANDed with each other and search.any)
                value:
                  search: { any: 'office', name: 'parole', address: 'atlanta', home_agency: ['8396b27b-95ef-45cb-9b4c-1132812a33e1', '5c2e9a4d-1f3b-4d7e-8a6c-2b9f0e1d3c4a'], exclude_from_colocation: [true] }
              searchSortAndPaginate:
                summary: Combine search, sort, and pagination
                value:
                  search: { any: 'springfield' }
                  pagination: { limit: 25, page: 2 }
                  data: { sort_by: address, order: desc, include_subagency: true }
      parameters:
        - name: client_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter to a specific client. Omit to return both agency-wide and client-specific fences.
        - name: q
          in: query
          schema:
            type: string
          description: |
            Plain, case-insensitive substring search, applied in memory
            after the DB fetch (agency/client/status filters are still
            pushed to the query — only search/sort/pagination happen
            application-side) rather than as a SQL/Oracle Text clause. No
            prefix syntax or boolean operators — a match against *any* of
            the fields below is enough.

            | Field | Matches |
            |---|---|
            | name | The geofence's name. |
            | address | The geo_json address object, flattened to text (street, city, state/province, postal code). |
            | exclude_from_colocation | Matched as plain text against `"true"`/`"false"` — not a boolean filter. `q=true` matches every excluded geofence, `q=false` every included one. |

            Coordinates and radius are not searchable — they're numeric,
            not text, and rarely known/typed by a caller. A query shorter
            than 3 characters is treated as invalid and matches nothing
            (returns an empty page, not a 400) — the same minimum
            sso-agency-service's agency search uses. The owning (home)
            agency isn't text-searchable — filter it by UUID with
            home_agency instead. Examples: `Downtown`, `Main St`, `true`.
          examples:
            name:
              summary: Search by name
              value: 'Downtown'
            address:
              summary: Search by address
              value: 'Main St'
            excludedFromColocation:
              summary: Only geofences excluded from co-location
              value: 'true'
        - name: name
          in: query
          schema:
            type: string
          description: |
            Per-field filter. Unlike q (which matches if *any* field
            contains it), every per-field filter supplied — name, address,
            home_agency, exclude_from_colocation — must match (AND), and
            they AND with q as well. name/address are case-insensitive
            substring matches against the same text q searches (address =
            the flattened geo_json address); there is no 3-character
            minimum for these. home_agency and exclude_from_colocation
            accept several values (see their own descriptions). Omit a
            param to not filter on that field. Example:
            `?name=parole&address=atlanta&home_agency=8396b27b-95ef-45cb-9b4c-1132812a33e1,5c2e9a4d-1f3b-4d7e-8a6c-2b9f0e1d3c4a&exclude_from_colocation=true`.
          example: 'Parole'
        - name: address
          in: query
          schema:
            type: string
          description: >
            Case-insensitive substring filter on the geofence's address —
            the geo_json address object flattened to text (street, city,
            state/province, postal code, country), the same text q
            searches. ANDed with the other filters and q; no 3-character
            minimum. Geofences with no address never match a non-blank
            value.
          example: 'Atlanta'
        - name: home_agency
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
              format: uuid
          description: >
            One or more agency UUIDs, either repeated
            (`?home_agency=<uuid1>&home_agency=<uuid2>`) or comma-separated
            (`?home_agency=<uuid1>,<uuid2>`). A geofence's home agency is
            the agency that owns it (its agency_id), so this matches
            geofences owned by *any* of the given agencies (OR within this
            field), ANDed with the other filters and q. It only narrows the
            agency scope already being listed (the path agency, plus
            subagencies with includeSubagency) — it never widens it. A
            non-UUID entry is a 400.
          example: ['8396b27b-95ef-45cb-9b4c-1132812a33e1', '5c2e9a4d-1f3b-4d7e-8a6c-2b9f0e1d3c4a']
        - name: exclude_from_colocation
          in: query
          style: form
          explode: true
          schema:
            type: array
            items:
              type: boolean
          description: >
            One or more booleans, repeated or comma-separated, OR within
            this field and ANDed with the other filters and q: `true`
            returns only geofences excluded from co-location, `false` only
            included ones, `true,false` both (same as omitting it). Only
            `true`/`false` are accepted (case-insensitive) — any other entry,
            including shorthands like `1`, `0`, `t` or `f`, is a 400.
          example: [true]
        - name: includeSubagency
          in: query
          schema:
            type: boolean
            default: false
          description: >
            Also list the path agency's subagencies — resolved live via the
            agency service, limited to the ones the caller is allowed to see
            (all of them for System/Global/Read-Only Administrators,
            otherwise only associated agencies; the whole hierarchy when JWT
            verification is disabled, a dev-only setting). Use home_agency
            to narrow to specific ones. If the agency service call fails,
            the request fails with 502 `ERR_AGENCY_SERVICE`.
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, archived, deleted]
          description: Omit to return every non-deleted geofence (deleted rows only surface when explicitly requested).
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 25
            maximum: 200
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [name, exclude_from_colocation, address]
          description: Field to sort by, applied in memory after the DB fetch (see the q parameter above) — exactly the same fields q searches. address sorts by the geofence's street address; geofences with no address sort last regardless of direction. Coordinates/radius aren't sortable (same reason they aren't searchable). Omit to keep the default (most recently created first).
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort direction for sort_by. Anything other than "desc" is treated as "asc".
      responses:
        '200':
          description: Paginated geofence list.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeofenceListResponse'
              examples:
                nameOrAddressMatch:
                  summary: A Circle geofence matching a name/address search
                  value:
                    data:
                      - id: c1a1a7f0-1234-4a2b-9c3d-000000000001
                        agency_id: 8396b27b-95ef-45cb-9b4c-1132812a33e1
                        name: Downtown Office
                        geo_json:
                          type: Circle
                          coordinates: [-84.388, 33.749]
                          radius: 500
                          address: { street1: 123 Main St, city: Atlanta }
                        status: active
                        user_id: d2b2b8f1-2345-4a2b-9c3d-000000000002
                        exclude_from_colocation: false
                        version: 1
                    meta: { limit: 25, page: 1, pages: 1, total: 1 }
                excludedMatch:
                  summary: A Polygon geofence matching q=true (only geofences excluded from co-location)
                  value:
                    data:
                      - id: c1a1a7f0-1234-4a2b-9c3d-000000000003
                        agency_id: 8396b27b-95ef-45cb-9b4c-1132812a33e1
                        name: Parole Office
                        geo_json:
                          type: Polygon
                          coordinates: [[[-84.4, 33.75], [-84.39, 33.75], [-84.39, 33.76], [-84.4, 33.76], [-84.4, 33.75]]]
                        status: active
                        user_id: d2b2b8f1-2345-4a2b-9c3d-000000000002
                        exclude_from_colocation: true
                        version: 2
                    meta: { limit: 25, page: 1, pages: 1, total: 1 }
        '400':
          $ref: '#/components/responses/BadRequest'
        '502':
          $ref: '#/components/responses/BadGateway'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    AgencyId:
      name: agency_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Agency scoping every geofence route.
    CorrelationId:
      name: x-correlation-id
      in: header
      required: false
      schema:
        type: string
      description: Propagated through to logs via log.CorrelationIdMiddleware; the caller may supply one, otherwise it is simply absent from log correlation.
  responses:
    BadGateway:
      description: An upstream agency-service call failed (e.g. the includeSubagency hierarchy lookup) — code ERR_AGENCY_SERVICE, input agency_id.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Invalid request (bad UUID, failed JSON binding, invalid geometry, etc).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalError:
      description: Unexpected server/database error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Geofence:
      allOf:
        - $ref: '#/components/schemas/GeofenceRequest'
        - type: object
          required:
            - id
            - version
            - agency_id
          properties:
            id:
              type: string
              format: uuid
            agency_id:
              type: string
              format: uuid
              description: Agency this geofence belongs to.
            updated_by:
              type: string
              format: uuid
              nullable: true
              description: Added beyond the original fragment. Identity of whoever last updated this record, if it's been updated; null otherwise.
            version:
              type: integer
              description: Version number of the geofence record, incremented on each update.
    GeofenceListRequest:
      type: object
      description: >
        Optional LIST-body envelope, mirroring sso-user-service's
        UserListRequest, for `LIST /agencies/{agency_id}/geofences`. Detected
        by the presence of a top-level search, pagination, or data key; if
        none of those three keys are present — including an empty or
        non-JSON body — every field falls back to its query-string
        equivalent instead. When the body is used, every query param is
        ignored except client_id (no body field; always read from the query
        string).
      properties:
        search:
          type: object
          description: >
            Optional search controls: any is the free-text search (equivalent
            to the q query param, matches if any field contains it); name,
            address, home_agency and exclude_from_colocation are per-field
            filters (equivalent to the same-named query params) that must
            all match (AND), together with any.
          properties:
            any:
              type: string
              description: |
                Plain, case-insensitive substring search, equivalent to the
                q query param — applied in memory after the DB fetch, not
                as a SQL/Oracle Text clause. No prefix syntax or boolean
                operators; a match against *any* of the fields below is
                enough.

                | Field | Matches |
                |---|---|
                | name | The geofence's name. |
                | address | The geo_json address object, flattened to text (street, city, state/province, postal code). |
                | exclude_from_colocation | Matched as plain text against `"true"`/`"false"` — not a boolean filter. `true` matches every excluded geofence, `false` every included one. |

                Coordinates and radius are not searchable. A value shorter
                than 3 characters is invalid and matches nothing (an empty
                page, not a 400). Example: `Downtown`.
              example: 'Downtown'
            name:
              type: string
              description: >
                Equivalent to the name query param — case-insensitive
                substring filter on the name, ANDed with the other search
                fields and any.
              example: 'Parole'
            address:
              type: string
              description: >
                Equivalent to the address query param — case-insensitive
                substring filter on the flattened geo_json address, ANDed
                with the other search fields and any.
              example: 'Atlanta'
            home_agency:
              oneOf:
                - type: string
                - type: array
                  items:
                    type: string
                    format: uuid
              description: >
                Equivalent to the home_agency query param — a single agency
                UUID or an array of them. Matches geofences owned by *any*
                of these agencies (OR within this field), ANDed with the
                other search fields and any; only narrows the listed agency
                scope, never widens it.
              example: ['8396b27b-95ef-45cb-9b4c-1132812a33e1', '5c2e9a4d-1f3b-4d7e-8a6c-2b9f0e1d3c4a']
            exclude_from_colocation:
              oneOf:
                - type: boolean
                - type: array
                  items:
                    type: boolean
              description: >
                Equivalent to the exclude_from_colocation query param — a
                single boolean or an array (OR within this field, so
                [true, false] matches both), ANDed with the other search
                fields and any. Omit it, or send null or [], to return both.
                A null array entry is a 400.
              example: [true, false]
        pagination:
          type: object
          description: Optional paging controls.
          properties:
            limit:
              type: integer
              minimum: 1
              maximum: 200
              default: 25
              description: Page size, equivalent to the page_size query param.
              example: 25
            page:
              type: integer
              minimum: 1
              default: 1
              description: 1-based page number, equivalent to the page query param.
              example: 1
        data:
          type: object
          description: Geofence-list-specific filters and sort.
          properties:
            client_id:
              type: string
              format: uuid
              deprecated: true
              description: >
                Not supported in the request body — this field is ignored.
                Use the client_id query parameter to filter to a specific
                client; it works regardless of whether search/pagination/data
                are also supplied in the body.
            status:
              type: string
              enum: [active, inactive, archived, deleted]
              description: >
                Equivalent to the status query param. Omit to return every
                non-deleted geofence — deleted rows only surface when
                explicitly requested.
              example: active
            sort_by:
              type: string
              enum: [name, exclude_from_colocation, address]
              description: >
                Equivalent to the sort_by query param, applied in memory
                after the DB fetch. address sorts by the
                geofence's street address (geofences with no address sort
                last, regardless of direction). Coordinates/radius aren't
                sortable (same reason they aren't searchable). Omit to
                keep the default order (most recently created first).
              example: name
            order:
              type: string
              enum: [asc, desc]
              default: asc
              description: >
                Sort direction for sort_by, equivalent to the order query
                param. Anything other than "desc" is treated as "asc".
              example: desc
            include_subagency:
              type: boolean
              default: false
              description: >
                Equivalent to the includeSubagency query param: include
                subagencies (resolved via the agency service, keyed off the
                agency_id path param) in the list query.
              example: true
    GeofenceListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Geofence'
        meta:
          type: object
          properties:
            limit:
              type: integer
            page:
              type: integer
            pages:
              type: integer
            total:
              type: integer
    ErrorDetail:
      type: object
      properties:
        code:
          type: string
        input:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'